# use atomic operations instead of mutexes in Counter and Gauge