This is the first of a couple of PRs I'm going to submit to attempt to make metric collection a bit faster.
While profiling production software I noticed that metric collection was taking an important part of the time spent, there were a couple of short wins like getting rid of some unused metrics or increasing the datadog client buffer size but now I'm down to having to optimize the code itself.

This PR changes the implementation of Counter and Gauge to use atomic operations instead of mutexes, which trims ~15% of time spent in the Counter and Gauge methods.
On top of being a quicker operation to execute this is also the kind of change which usually helps programs achieve higher throughput as there's no contention point anymore where goroutines may be paused.

Here are numbers from the benchmarks:
```
benchmark                   old ns/op     new ns/op     delta
BenchmarkCounter/Incr-4     64.4          54.8          -14.91%
BenchmarkCounter/Add-4      61.7          52.1          -15.56%
BenchmarkCounter/Set-4      59.3          52.7          -11.13%
BenchmarkGauge/Incr-4       64.7          55.0          -14.99%
BenchmarkGauge/Decr-4       62.3          55.2          -11.40%
BenchmarkGauge/Add-4        63.1          53.2          -15.69%
BenchmarkGauge/Set-4        62.8          52.5          -16.40%
```

Please take a look and let me know if you have any concerns.