@abraithwaite 2017-07-10
Nice.  Although, what is the reason why we're keeping a copy of the metrics in memory in the stats package if it's just being sent to Datadog/Prometheus clients?