@achille-roussel 2017-07-10
The Counter/Gauge/Histogram/Timer types are used as a way to construct metrics, it prevents the need to know te metric name and tags at the place where they are used, as well as expose extra operations (like Counter.Set), which require maintaining state. 

I have more optimisations in mind that will leverage those types as well.